+++ src/main/java/com/web/FileTran/dao/filesMapper.java
@@ -9,8 +9,8 @@ public interface filesMapper {
     // TODO 需要修改存储过程,添加错误信息输出
     /**
      * 创建(上传)文件
-     * @param params 输入 long parentFolderId,long creatorId,String fileName,Blob fileContent
-     * @return 输出 long newFileId,String errorMessage
+     * @param params 输入 int parentFolderId,int creatorId,String fileName,Blob fileContent
+     * @return 输出 int newFileId,String errorMessage
      */
     @MapKey("id")
     Map<String, Object> createFile(Map<String, Object> params);
@@ -18,8 +18,8 @@ public interface filesMapper {
     // TODO 需要修改存储过程,添加错误信息输出
     /**
      * 创建快捷方式
-     * @param params 输入 long parentFolderId,long creatorId,String fileName,long targetFileId
-     * @return 输出 long newShortcutId,String errorMessage
+     * @param params 输入 int parentFolderId,int creatorId,String fileName,int targetFileId
+     * @return 输出 int newShortcutId,String errorMessage
      */
     @MapKey("id")
     Map<String, Object> createFileShortcut(Map<String, Object> params);
@@ -29,7 +29,7 @@ public interface filesMapper {
      * 删除文件
      * @param fileId 待删除文件id
      */
-    void deleteFile(long fileId);
+    void deleteFile(int fileId);
 
     // TODO 需要修改存储过程,添加错误信息输出
     /**
@@ -38,7 +38,7 @@ public interface filesMapper {
      * @return 输出 Blob fileBlob,String errorMessage
      */
     @MapKey("id")
-    Map<String, Object> downloadFile(long fileId);
+    Map<String, Object> downloadFile(int fileId);
 
     /**
      * 获得文件信息
@@ -46,12 +46,12 @@ public interface filesMapper {
      * @return pojo类
      */
     @MapKey("id")
-    files getFileInfo(long fileId);
+    files getFileInfo(int fileId);
 
     /**
      * 查询文件是否存在
      * @param fileId 文件id
      * @return 文件是否存在
      */
-    boolean checkFileExists(long fileId);
+    boolean checkFileExists(int fileId);
 }
